requirements/base.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fastapi==0.115.12
uvicorn[standard]==0.34.0
psycopg[binary]==3.2.3
python-decouple==3.8
dj-database-url==2.2.0
django-cors-headers==4.4.0
whitenoise==6.8.2
pydantic==2.10.6
asgiref==3.8.1
